Refactor MyEvaluator into its own file
Moves the `MyEvaluator` class from `mydsl/__init__.py` to a new file `mydsl/evaluator.py` and updates import statements in `tests/test_string_literals.py`.

- Empties `mydsl/__init__.py` and adds a comment indicating it as a package initializer.
- Creates `mydsl/evaluator.py` with the `MyEvaluator` class and its methods, ensuring to import `Transformer` from `lark` at the top of the file.
- Updates `tests/test_string_literals.py` to import `MyEvaluator` from `mydsl.evaluator` instead of `mydsl`, aligning with the new file structure.


---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/andreisavu/lark-escaping/pull/20?shareId=e3836219-725e-4ef6-bdb8-263b4e6acb34).
